requirements.txt
Flask
Flask-MySQLdb
python-dotenv
Flask-CORS
schedule